restrict most stations, sensors and sensordata routes to the owner of the resource

+++ src/stations/stations.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Station } from './stations.entity';
@@ -42,16 +42,27 @@ export class StationsService {
     return this.stationRepo.save(station);
   }
 
-  async updateStationById(stationId: string, updateStationDto: UpdateStationDto) {
+  async updateStationById(stationId: string, updateStationDto: UpdateStationDto, user: User) {
     const station = await this.findStationById(stationId);
-    Object.assign(station, updateStationDto);
+    if (this.stationBelongsToCurrentUser(station, user.id)) {
+      Object.assign(station, updateStationDto);
 
-    return this.stationRepo.save(station);
+      return this.stationRepo.save(station);
+    }
   }
 
-  async removeStationById(stationId: string) {
+  async removeStationById(stationId: string, user: User) {
     const station = await this.findStationById(stationId);
+    if (this.stationBelongsToCurrentUser(station, user.id)) {
+      return this.stationRepo.remove(station);
+    }
+  }
 
-    return this.stationRepo.remove(station);
+  async stationBelongsToCurrentUser(station: Station, userId: string) {
+    if (station.userId !== userId) {
+      throw new ForbiddenException('User is not the owner of that station');
+    }
+
+    return true;
   }
 }
